% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlternateLasso.R, R/plot.R, R/print.R,
%   R/summary.R
\name{AlternateLasso}
\alias{AlternateLasso}
\alias{plot.AlternateLasso}
\alias{print.AlternateLasso}
\alias{summary.AlternateLasso}
\title{Alternate Lasso Feature Selection}
\usage{
AlternateLasso(X, y, model = NULL, rho = 0.1, featurename = NULL,
  verbose = TRUE, ...)

\method{plot}{AlternateLasso}(obj, ...)

\method{print}{AlternateLasso}(obj)

\method{summary}{AlternateLasso}(obj)
}
\arguments{
\item{X}{[matrix] Desgin matrix}

\item{y}{[vector] Response valuables}

\item{model}{\code{\link{glmnet}} class object.}

\item{rho}{[numeric] Cut-off value. See \code{lambda} of \code{\link{glmnet}}}

\item{featurename}{[character vector] If NULL, set colnames(X)}

\item{verbose}{Show message during fit. default is \code{TRUE.}}

\item{...}{Other options path to \code{\link{pforeach}}}

\item{obj}{"AlternateLasso" class object.}

\item{...}{adjuster options path to \code{networkD3::sankeyNetwork}.}

\item{obj}{"AlternateLasso" class object.}
}
\value{
"AlternateLasso" class object. A list of original model and alternative features.
}
\description{
R package for the alternate features search proposed in the following paper.
S. Hara, T. Maehara, Finding Alternate Features in Lasso, arXiv:1611.05940, 2016.
Complete and original Python code is available from https://github.com/sato9hara/LassoVariants
}
