% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{convert.graphDataFrame}
\alias{convert.graphDataFrame}
\title{convert from Alternate Lasso object to data.frame}
\usage{
convert.graphDataFrame(obj)
}
\description{
convert from Alternate Lasso object to data.frame
}
