% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%do\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{foreach}{\code{\link[foreach]{\%do\%}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

